Refactored directory layout
Now the directory layout is Go friendly
