+++ .travis.yml
@@ -4,14 +4,11 @@ go:
   - 1.13.1
   - master
 
-before_script:
-- "go get github.com/google/go-cmp/cmp"
-- "go get github.com/gorilla/context"
-- "go get github.com/gorilla/mux"
-- "go get golang.org/x/net/context/ctxhttp"
+go_import_path:
+  - github.com/gorilla/mux
+  - github.com/google/go-cmp/cmp
 
 script:
-- "pushd ./API"
-- "pushd ./MXLookup"
-- "go test -v"
-- "popd"
+  - "pushd ./API"
+  - go test -v ./...
+  - popd
