+++ main.go
@@ -5,11 +5,10 @@ import (
 	"log"
 	"net/http"
 	"github.com/gorilla/mux"
-
-	"./MXLookup"
-	"./AddrLookup"
-	"./CNAMELookup"
-	"./HostLookup"
+	"github.com/HauptJ/Golang-API-DNS/API/MXLookup"
+	"github.com/HauptJ/Golang-API-DNS/API/AddrLookup"
+	"github.com/HauptJ/Golang-API-DNS/API/CNAMELookup"
+	"github.com/HauptJ/Golang-API-DNS/API/HostLookup"
 )
